+++ scout/server/links.py
@@ -1,6 +1,7 @@
 from pprint import pprint as pp
 
 from scout.constants import SPIDEX_HUMAN
+from scout.utils.convert import amino_acid_residue_change_3_to_1
 
 
 def add_gene_links(gene_obj, build=37):
@@ -56,6 +57,23 @@ def add_gene_links(gene_obj, build=37):
     # Add links that use ucsc link
     gene_obj["ucsc_link"] = ucsc(gene_obj.get("ucsc_id"))
     gene_obj["genemania_link"] = genemania(hgnc_symbol)
+    gene_obj["oncokb_link"] = oncokb(hgnc_symbol)
+    gene_obj["cbioportal_link"] = cbioportal_gene(hgnc_symbol)
+    gene_obj["iarctp53_link"] = iarctp53(hgnc_symbol)
+
+
+def cbioportal_gene(hgnc_symbol):
+    link = "https://www.cbioportal.org/ln?q={}%3AMUT"
+    if not hgnc_symbol:
+        return None
+    return link.format(hgnc_symbol)
+
+
+def oncokb(hgnc_symbol):
+    link = "https://www.oncokb.org/gene/{}"
+    if not hgnc_symbol:
+        return None
+    return link.format(hgnc_symbol)
 
 
 def genemania(hgnc_symbol):
@@ -197,7 +215,7 @@ def ucsc(ucsc_id):
     return link.format(ucsc_id)
 
 
-def add_tx_links(tx_obj, build=37):
+def add_tx_links(tx_obj, build=37, hgnc_symbol=None):
     try:
         build = int(build)
     except ValueError:
@@ -225,6 +243,9 @@ def add_tx_links(tx_obj, build=37):
     tx_obj["varsome_link"] = varsome(
         build, tx_obj.get("refseq_id"), tx_obj.get("coding_sequence_name")
     )
+    tx_obj["tp53_link"] = mutantp53(tx_obj.get("hgnc_id"), tx_obj.get("protein_sequence_name"))
+    tx_obj["cbioportal_link"] = cbioportal(hgnc_symbol, tx_obj.get("protein_sequence_name"))
+    tx_obj["mycancergenome_link"] = mycancergenome(hgnc_symbol, tx_obj.get("protein_sequence_name"))
 
     return tx_obj
 
@@ -300,6 +321,16 @@ def varsome(build, refseq_id, protein_sequence_name):
     link = "https://varsome.com/variant/hg{}/{}:{}"
 
     return link.format(build, refseq_id, protein_sequence_name)
+
+
+def iarctp53(hgnc_symbol):
+
+    if hgnc_symbol != "TP53":
+        return None
+
+    link = "https://p53.iarc.fr/TP53GeneVariations.aspx"
+
+    return link
 
 
 ############# Variant links ####################
@@ -445,6 +476,47 @@ def ucsc_link(variant_obj, build=None):
         )
 
     return url_template.format(this=variant_obj)
+
+
+def mycancergenome(hgnc_symbol, protein_sequence_name):
+    link = "https://www.mycancergenome.org/content/alteration/{}-{}"
+
+    if not hgnc_symbol:
+        return None
+    if not protein_sequence_name:
+        return None
+
+    protein_change = amino_acid_residue_change_3_to_1(protein_sequence_name)
+
+    if not protein_change:
+        return None
+
+    return link.format(hgnc_symbol, protein_change.lower())
+
+
+def cbioportal(hgnc_symbol, protein_sequence_name):
+    link = "https://www.cbioportal.org/ln?q={}:MUT%20%3D{}"
+
+    if not hgnc_symbol:
+        return None
+    if not protein_sequence_name:
+        return None
+
+    protein_change = amino_acid_residue_change_3_to_1(protein_sequence_name)
+
+    if not protein_change:
+        return None
+
+    return link.format(hgnc_symbol, protein_change)
+
+
+def mutantp53(hgnc_id, protein_variant):
+    if hgnc_id != 11998:
+        return None
+
+    url_template = "http://mutantp53.broadinstitute.org/?query={}"
+
+    return url_template.format(protein_variant)
 
 
 def alamut_link(variant_obj, build=None):