+++ CHANGELOG.md
@@ -8,6 +8,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 ### Added
 - Extend the delete variants command to print analysis date, track, institute, status and research status
 - Delete variants by type of analysis (wgs|wes|panel)
+- Links to cBioPortal, MutanTP53, IARC TP53, OncoKB, MyCancerGenome
 ### Fixed
 ### Changed
 - Print output of variants delete command as a tab separated table